backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "universe-visualizer-backend"
version = "0.1.0"
description = "FastAPI backend for the Universe Visualizer observatory platform."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "pydantic-settings>=2.5.0,<3.0.0",
  "sgp4>=2.24",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
